refactor: extract swebench dataset ownership
